+++ src/main/java/pl/plajer/pinata/commands/MainCommand.java
