+++ src/main/java/pl/plajer/pinata/creator/ChatReaction.java
