+++ src/main/java/pl/plajer/pinata/creator/SelectorEvents.java
