+++ src/main/java/pl/plajer/pinata/creator/SelectorInventories.java
