+++ src/main/java/pl/plajer/pinata/handlers/PinataManager.java
