+++ src/main/java/pl/plajer/pinata/handlers/SignManager.java
