+++ src/main/java/pl/plajer/pinata/pinata/LivingPinata.java
