+++ src/main/resources/plugin.yml
@@ -1,5 +1,5 @@
 name: Pinata
-version: 3.0.0b
+version: 3.0.0c
 author: Plajer
 description: Awesome pinata sheep mobs for your server!
 website: https://www.spigotmc.org/resources/pi%C3%B1ata-1-8-1-12.46655/
